review: move to remote_process_impl.h/cc
