+++ source/client/process_impl.cc
@@ -1,6 +1,5 @@
 #include "client/process_impl.h"
 
-#include <grpc++/grpc++.h>
 #include <sys/file.h>
 
 #include <chrono>
@@ -37,7 +36,6 @@
 
 #include "api/client/options.pb.h"
 #include "api/client/output.pb.h"
-#include "api/client/service.grpc.pb.h"
 
 #include "common/frequency.h"
 #include "common/uri_impl.h"
@@ -356,7 +354,6 @@ bool ProcessImpl::run(OutputCollector& collector) {
   } catch (const UriException&) {
     return false;
   }
-
   int number_of_workers = determineConcurrency();
   shutdown_ = false;
   const std::vector<ClientWorkerPtr>& workers = createWorkers(number_of_workers);
@@ -434,54 +431,6 @@ void ProcessImpl::setupForHRTimers() {
   // Getting to libevent for the other approach is going to introduce more code as we would need to
   // derive our own customized versions of certain Envoy concepts.
   putenv(const_cast<char*>("EVENT_PRECISE_TIMER=1"));
-}
-
-RemoteProcessImpl::RemoteProcessImpl(const Options& options) : options_(options) {}
-
-bool RemoteProcessImpl::run(OutputCollector& collector) {
-  UriPtr uri;
-
-  try {
-    uri = std::make_unique<UriImpl>(options_.nighthawkService());
-  } catch (const UriException&) {
-    ENVOY_LOG(error, "Bad service uri: {}", options_.nighthawkService());
-    return false;
-  }
-
-  auto channel = grpc::CreateChannel(fmt::format("{}:{}", uri->hostWithoutPort(), uri->port()),
-                                     grpc::InsecureChannelCredentials());
-  auto stub = std::make_unique<nighthawk::client::NighthawkService::Stub>(channel);
-  grpc::ClientContext context;
-  nighthawk::client::ExecutionRequest request;
-  nighthawk::client::ExecutionResponse response;
-  auto r = stub->ExecutionStream(&context);
-
-  *request.mutable_start_request()->mutable_options() = *options_.toCommandLineOptions();
-  // We don't forward the option that requests remote execution. Today,
-  // nighthawk_service will ignore the option, but if someone ever changes that this
-  // is probably desireable.
-  request.mutable_start_request()->mutable_options()->mutable_nighthawk_service()->Clear();
-
-  if (r->Write(request, {}) && r->Read(&response)) {
-    if (response.has_output()) {
-      collector.setOutput(response.output());
-    } else {
-      ENVOY_LOG(error, "Remote execution failed");
-    }
-    if (response.has_error_detail()) {
-      ENVOY_LOG(error, "have error detail: {}", response.error_detail().DebugString());
-    }
-    if (!r->WritesDone()) {
-      ENVOY_LOG(warn, "writeDone() failed");
-    } else {
-      auto status = r->Finish();
-      return status.ok();
-    }
-  } else {
-    ENVOY_LOG(error, "Failure while communicating with the remote service");
-  }
-
-  return false;
 }
 
 } // namespace Client
