+++ source/client/process_impl.h
@@ -107,18 +107,5 @@ private:
   bool shutdown_{true};
 };
 
-/**
- * Will delegate execution to a remote nighthawk_service using gRPC.
- */
-class RemoteProcessImpl : public Process, public Envoy::Logger::Loggable<Envoy::Logger::Id::main> {
-public:
-  RemoteProcessImpl(const Options& options);
-  bool run(OutputCollector& collector) override;
-  void shutdown() override{};
-
-private:
-  const Options& options_;
-};
-
 } // namespace Client
 } // namespace Nighthawk
